refactor(sensor): split aura_telemetry_native.cpp into domain-focused files
Split the 1516-line monolith into 10 focused files (telemetry_utils,
collect_cpu, collect_processes, collect_process_details, collect_disk,
collect_network, collect_gpu, collect_thermal, process_management).
Internal telemetry_utils.h provides shared helpers across modules.
No behavioral changes — pure file reorganization.
